feat: 目标监控数据增加删除二次确认

+++ MY_TargetMon/src/MY_TargetMon_CP.lua
@@ -530,8 +530,10 @@ function D.Open(szConfigUUID)
 		color = { 255, 0, 0 },
 		buttonStyle = 'FLAT',
 		onClick = function()
-			MY_TargetMon.DeleteConfig(szConfigUUID)
-			MY_TargetMon_CP.Close()
+			X.Confirm(_L['Sure to delete config? This operation can not be undone.'], function()
+				MY_TargetMon.DeleteConfig(szConfigUUID)
+				MY_TargetMon_CP.Close()
+			end)
 		end,
 	})
